feat: byte-addressable memory model (uint8_t* mem)
Change prog.mem from int64_t* to uint8_t* for byte-addressable memory.
Remove shl rax,3 from JIT LOAD/STORE (addresses are now byte offsets).
Multiply all address computations by 8 (cell index -> byte address).
Struct member offsets returned as byte offsets (cell * 8).
Array stride computed from element size in bytes.
Static local brace init routed through brace path with guard check.

Score: 1,212 -> 1,196 (regressions in alignment-sensitive tests).
Infrastructure correct; remaining: 16-byte alignment, pointer arith.

+++ test_gauntlet_runner.c
@@ -132,7 +132,7 @@ int main(int argc, char **argv) {
             wubu_mir_prog_t prog;
             if (hd_build_mir(src, &prog) != 0) { printf("build failed\n"); return 1; }
             /* Allocate mem like hd_run_prog does (JIT needs it for LOAD/STORE) */
-            int64_t *mem_ptr = prog.mem;
+            uint8_t *mem_ptr = prog.mem;
             if (mem_ptr == NULL) {
                 int64_t mem_hi = prog.total_mem;
                 if ((int64_t)(prog.next_vr_hi) - 1 > mem_hi) mem_hi = (int64_t)(prog.next_vr_hi) - 1;
